+++ packages/core/src/api/neon-proxy-rpc.ts
@@ -1,4 +1,4 @@
-import { GasToken, NeonEmulate, NeonProgramStatus, RPCResponse } from '../models';
+import { GasToken, GasTokenV2, NeonEmulate, NeonProgramStatusV2, NeonProgramStatus, RPCResponse } from '../models';
 
 export class NeonProxyRpcApi {
   neonProxyRpcUrl = '';
@@ -27,15 +27,45 @@ export class NeonProxyRpcApi {
     return this.proxy<NeonEmulate>('neon_emulate', params).then(d => d.result);
   }
 
-  async evmParams(): Promise<NeonProgramStatus> {
-    return this.proxy<NeonProgramStatus>('neon_getEvmParams', []).then(d => d.result);
+  async evmParams(): Promise<Partial<NeonProgramStatus>> {
+    return this.proxy<NeonProgramStatusV2>('neon_getEvmParams', []).then(d => {
+      //Use data mapper
+      return {
+        NEON_ACCOUNT_SEED_VERSION: d.result.neonAccountSeedVersion,
+        NEON_EVM_ID: d.result.neonEvmProgramId,
+        NEON_EVM_STEPS_LAST_ITERATION_MAX: d.result.neonMaxEvmStepsInLastIteration,
+        NEON_EVM_STEPS_MIN: d.result.neonMinEvmStepsInIteration,
+        NEON_GAS_LIMIT_MULTIPLIER_NO_CHAINID: d.result.neonGasLimitMultiplierWithoutChainId,
+        NEON_HOLDER_MSG_SIZE: d.result.neonHolderMessageSize,
+        NEON_PAYMENT_TO_TREASURE: d.result.neonPaymentToTreasury,
+        NEON_STORAGE_ENTRIES_IN_CONTRACT_ACCOUNT: d.result.neonStorageEntriesInContractAccount,
+        NEON_TREASURY_POOL_COUNT: d.result.neonTreasuryPoolCount,
+        NEON_TREASURY_POOL_SEED: d.result.neonTreasuryPoolSeed,
+      }
+    });
   }
 
   async gasTokenList(): Promise<GasToken[]> {
-    return this.proxy<GasToken[]>('neon_getGasTokenList', []).then(d => d.result);
+    return this.proxy<GasTokenV2[]>('neon_getGasTokenList', []).then(d => {
+      return d.result.map((d) => {
+        return {
+          tokenName: d.tokenName,
+          tokenMint: d.tokenMint,
+          tokenChainId: d.tokenChainID,
+        };
+      });
+    });
   }
 
   async nativeTokenList(): Promise<GasToken[]> {
-    return this.proxy<GasToken[]>('neon_getNativeTokenList', []).then(d => d.result);
+    return this.proxy<GasTokenV2[]>('neon_getNativeTokenList', []).then(d => {
+      return d.result.map((d) => {
+        return {
+          tokenName: d.tokenName,
+          tokenMint: d.tokenMint,
+          tokenChainId: d.tokenChainID,
+        };
+      });
+    });
   }
 }