+++ packages/web3/src/mint-transfer.ts
@@ -3,7 +3,7 @@ import {
   PublicKey,
   SystemProgram,
   Transaction,
-  TransactionInstruction
+  TransactionInstruction,
 } from '@solana/web3.js';
 import {
   Amount,
@@ -14,7 +14,7 @@ import {
   NeonProxyRpcApi,
   neonTransferMintTransaction,
   SPLToken,
-  toFullAmount
+  toFullAmount,
 } from '@neonevm/token-transfer-core';
 import {
   createSyncNativeInstruction,
@@ -42,7 +42,7 @@ export async function neonTransferMintTransactionWeb3(connection: Connection, pr
   const {
     neonKeys,
     legacyAccounts
-  } = await createClaimInstruction<SignTransactionResult>(proxyApi, signedTransaction);
+  } = await createClaimInstruction<SignTransactionResult>({ proxyApi, neonTransaction: signedTransaction, connection, neonEvmProgram, splToken, associatedTokenAddress, signerAddress: walletSigner.address, fullAmount });
   return neonTransferMintTransaction<Web3Account, SignTransactionResult>(connection, neonEvmProgram, solanaWallet, neonWallet, walletSigner, neonKeys, legacyAccounts, signedTransaction, splToken, fullAmount, chainId, neonHeapFrame);
 }
 
@@ -70,7 +70,7 @@ export async function createWrapAndTransferSOLTransaction(connection: Connection
   const {
     neonKeys,
     legacyAccounts
-  } = await createClaimInstruction<SignTransactionResult>(proxyApi, signedTransaction);
+  } = await createClaimInstruction<SignTransactionResult>({ proxyApi, neonTransaction: signedTransaction, connection, neonEvmProgram, splToken, associatedTokenAddress, signerAddress: walletSigner.address, fullAmount });
   const mintTransaction = await neonTransferMintTransaction<Web3Account, SignTransactionResult>(connection, neonEvmProgram, solanaWallet, neonWallet, walletSigner, neonKeys, legacyAccounts, signedTransaction, splToken, fullAmount, chainId, neonHeapFrame);
 
   if (!wSOLAccount) {