+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Neon EVM Token transfer SDK Changelog
 All notable changes to this project will be documented in this file.
 
+## [4.1.0] (2024-XX-XX)
+### Added
+* Add solanaOverrides in simulation request
+
+### Refactor
+* Update 'neon_getEvmParams' response model
+* Update 'neon_getNativeTokenList' response model
+* Update 'neon_emulate' response model
+* Update 'balanceAccount' for the 'TransactionExecuteFromInstruction'
+
 ## [4.0.0] (2024-02-06)
 ### Added
 * Ether.js library support